add compaction method
