+++ core/src/executor/datafusion/mod.rs
@@ -38,11 +38,13 @@ const POSITION_DELETE_TABLE: &str = "position_delete_table";
 const EQUALITY_DELETE_TABLE: &str = "equality_delete_table";
 const DEFAULT_PREFIX: &str = "10";
 
+#[derive(Default)]
 pub struct DataFusionExecutor {}
 
 #[async_trait]
 impl CompactionExecutor for DataFusionExecutor {
     async fn rewrite_files(
+        &self,
         file_io: FileIO,
         schema: Arc<Schema>,
         input_file_scan_tasks: InputFileScanTasks,
@@ -161,7 +163,7 @@ impl CompactionExecutor for DataFusionExecutor {
             .iter()
             .map(|f| f.file_size_in_bytes())
             .sum();
-        stat.rewritten_files_count = rewritten_files_count as u32;
+        stat.rewritten_files_count = rewritten_files_count;
 
         Ok(CompactionResult {
             data_files: output_data_files,
@@ -465,160 +467,5 @@ impl DataFusionTaskContext {
         self.equality_delete_files
             .as_ref()
             .is_some_and(|v| !v.is_empty())
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use futures_async_stream::for_await;
-    use iceberg::Catalog;
-    use iceberg::scan::FileScanTask;
-    use iceberg::table::Table;
-    use iceberg::writer::file_writer::location_generator::DefaultLocationGenerator;
-    use iceberg::{TableIdent, io::FileIOBuilder, transaction::Transaction};
-    use iceberg_catalog_sql::{SqlBindStyle, SqlCatalog, SqlCatalogConfig};
-    use std::collections::HashMap;
-    use std::sync::Arc;
-
-    use crate::CompactionError;
-    use crate::executor::{CompactionResult, InputFileScanTasks};
-    use crate::{CompactionConfig, CompactionExecutor, executor::DataFusionExecutor};
-
-    async fn build_catalog() -> SqlCatalog {
-        let sql_lite_uri = "postgresql://xxhx:123456@localhost:5432/demo_iceberg";
-        let warehouse_location = "s3a://hummock001/iceberg-data".to_owned();
-        let config = SqlCatalogConfig::builder()
-            .uri(sql_lite_uri.to_owned())
-            .name("demo1".to_owned())
-            .warehouse_location(warehouse_location)
-            .file_io(
-                FileIOBuilder::new("s3a")
-                    .with_prop("s3.secret-access-key", "hummockadmin")
-                    .with_prop("s3.access-key-id", "hummockadmin")
-                    .with_prop("s3.endpoint", "http://127.0.0.1:9301")
-                    .with_prop("s3.region", "")
-                    .build()
-                    .unwrap(),
-            )
-            .sql_bind_style(SqlBindStyle::DollarNumeric)
-            .build();
-        SqlCatalog::new(config).await.unwrap()
-    }
-
-    async fn get_tasks_from_table(table: Table) -> Result<InputFileScanTasks, CompactionError> {
-        let snapshot_id = table.metadata().current_snapshot_id().unwrap();
-
-        let scan = table
-            .scan()
-            .snapshot_id(snapshot_id)
-            .with_delete_file_processing_enabled(true)
-            .build()?;
-        let file_scan_stream = scan.plan_files().await?;
-
-        let mut position_delete_files = HashMap::new();
-        let mut data_files = vec![];
-        let mut equality_delete_files = HashMap::new();
-
-        #[for_await]
-        for task in file_scan_stream {
-            let task: FileScanTask = task?;
-            match task.data_file_content {
-                iceberg::spec::DataContentType::Data => {
-                    for delete_task in task.deletes.iter() {
-                        match &delete_task.data_file_content {
-                            iceberg::spec::DataContentType::PositionDeletes => {
-                                let mut delete_task = delete_task.clone();
-                                delete_task.project_field_ids = vec![];
-                                position_delete_files
-                                    .insert(delete_task.data_file_path.clone(), delete_task);
-                            }
-                            iceberg::spec::DataContentType::EqualityDeletes => {
-                                let mut delete_task = delete_task.clone();
-                                delete_task.project_field_ids = delete_task.equality_ids.clone();
-                                equality_delete_files
-                                    .insert(delete_task.data_file_path.clone(), delete_task);
-                            }
-                            _ => {
-                                unreachable!()
-                            }
-                        }
-                    }
-                    data_files.push(task);
-                }
-                _ => {
-                    unreachable!()
-                }
-            }
-        }
-        Ok(InputFileScanTasks {
-            data_files,
-            position_delete_files: position_delete_files.into_values().collect(),
-            equality_delete_files: equality_delete_files.into_values().collect(),
-        })
-    }
-
-    #[tokio::test]
-    async fn test_compact() {
-        let catalog = build_catalog().await;
-        let table_id = TableIdent::from_strs(vec!["demo_db", "test_all_delete"]).unwrap();
-        let table = catalog.load_table(&table_id).await.unwrap();
-
-        let manifest_list = table
-            .metadata()
-            .current_snapshot()
-            .unwrap()
-            .load_manifest_list(table.file_io(), table.metadata())
-            .await
-            .unwrap();
-
-        let mut data_file = vec![];
-        let mut delete_file = vec![];
-        for manifest_file in manifest_list.entries() {
-            let a = manifest_file.load_manifest(table.file_io()).await.unwrap();
-            let (entry, _) = a.into_parts();
-            for i in entry {
-                match i.content_type() {
-                    iceberg::spec::DataContentType::Data => {
-                        data_file.push(i.data_file().clone());
-                    }
-                    iceberg::spec::DataContentType::EqualityDeletes => {
-                        delete_file.push(i.data_file().clone());
-                    }
-                    iceberg::spec::DataContentType::PositionDeletes => {
-                        delete_file.push(i.data_file().clone());
-                    }
-                }
-            }
-        }
-        let all_file_scan_tasks = get_tasks_from_table(table.clone()).await.unwrap();
-        let file_io = table.file_io().clone();
-        let schema = table.metadata().current_schema();
-        let default_location_generator =
-            DefaultLocationGenerator::new(table.metadata().clone()).unwrap();
-        let CompactionResult {
-            data_files: output_data_files,
-            stat: _,
-        } = DataFusionExecutor::rewrite_files(
-            file_io,
-            schema.clone(),
-            all_file_scan_tasks,
-            Arc::new(CompactionConfig {
-                batch_parallelism: Some(4),
-                target_partitions: Some(4),
-                data_file_prefix: None,
-            }),
-            default_location_generator.dir_path,
-        )
-        .await
-        .unwrap();
-        let txn = Transaction::new(&table);
-        let mut rewrite_action = txn.rewrite_files(None, vec![]).unwrap();
-        rewrite_action
-            .add_data_files(output_data_files.clone())
-            .unwrap();
-        rewrite_action.delete_files(data_file).unwrap();
-        rewrite_action.delete_files(delete_file).unwrap();
-        let tx = rewrite_action.apply().await.unwrap();
-        tx.commit(&catalog).await.unwrap();
     }
 }
