+++ core/src/lib.rs
@@ -1,6 +1,7 @@
 #![feature(proc_macro_hygiene, stmt_expr_attributes)]
 #![feature(coroutines)]
 
+pub mod compaction;
 pub mod config;
 pub mod error;
 pub mod executor;
